+++ docs/CHANGELOG.md
@@ -1,6 +1,6 @@
 # GoNB Changelog
 
-## Next
+## v0.11.2, 2025/10.24 Support for various OSes; Fixed logging; Increased gopls timeout and improved error handling.
 
 * Support Jupyter for more OSSes (thanks @yuce!)
 * Updated dependencies